+++ shorturl-backend-naive/src/main/java/com/example/shorturl/repository/ShortUrlRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ShortUrlRepository extends JpaRepository<ShortUrl, Integer> {
+public interface ShortUrlRepository extends JpaRepository<ShortUrl, Long> {
     Page<ShortUrl> findAllByUserId(Integer userId, Pageable pageable);
 }
