+++ shorturl-backend-naive/src/main/java/com/example/shorturl/service/ShortUrlService.java
@@ -5,5 +5,5 @@ import com.example.shorturl.dto.Message;
 public interface ShortUrlService {
     Message addToMyShortUrls(String url);
     Message findAllMyShortUrls(int page, int size);
-    Message deleteMyShortUrlById(int id);
+    Message deleteMyShortUrlById(long id);
 }
